Update generated code to recent elixir style
Along with making the code generated already formated
the way mix format does, some small changes were added
to speed up new code implementations:

moduledoc and doc were added to remind coders to fill
the main responsability and expected behavior in a way
credo would evaluate the code without raising warnings.

Altough pure elixir is still the supported, the fact is
most of the time the code implemented is added to a phoenix
codebase, for that reason DataCase is now the standard.